Use external telegraf-registry repo
this is for supporting external plugins.

external plugins will depend on a few telegraf interface types, as well
as a commong telegraf registry.

because the Go 'plugin' package asserts that commit hashes of shared
dependencies must match, we are breaking these shared dependencies into
a separate repo that should very rarely change, which will allow
cross-version plugins to work.

see #1717